still compiles

+++ mp_stickers/src/Image.cpp
@@ -2,11 +2,17 @@
 
 void Image::lighten() {}
 void Image::lighten(double amount) {}
+// void Image::darken() {}
 void Image::darken(double amount) {}
 void Image::saturate() {}
 void Image::saturate(double amount) {}
-void Image::rotateColor(int degrees) {}
+// void Image::desaturate() {}
+// void Image::desaturate(double amount) {}
+// void Image::grayscale();
+void Image::rotateColor(double degrees) {}
+// void Image::Illinify() {}
 void Image::scale(double factor) {}
+// void scale(unsigned w, unsigned h) {}
 
 //getters
 unsigned int Image::width() const { return width_; }
